insert Interface IVMScope; switch some interfaces to interfaces.ts
