+++ src/utils/utils.ts
@@ -1,48 +1,7 @@
 ﻿
 //#region Import
 import { logging } from "./logger";
-//#endregion
-
-//#region interfaces
-interface IPropertiesDefault {
-    ref: string | boolean;
-    defaultValue: string;
-}
-
-interface IDefinitionObject {
-    definition: any;
-    getDefinition: any;
-}
-
-export interface IRegisterDirective {
-    directive(name: string, directiveFactory: ng.Injectable<ng.IDirectiveFactory>): void;
-    directive(object: { [directiveName: string]: ng.Injectable<ng.IDirectiveFactory> }): void;
-    filter(name: string, filterFactoryFunction: ng.Injectable<Function>): ng.IModule;
-    filter(object: { [name: string]: ng.Injectable<Function> }): ng.IModule;
-}
-
-
-export interface ICalcCubeElement {
-    qState: string;
-    cId: string;
-    qGroupFieldDefs: Array<string>;
-    qFallbackTitle: string;
-}
-
-export interface IMenuElement {
-    buttonType: string;
-    isVisible: boolean;
-    isEnabled: boolean;
-    isChecked?: boolean;
-    icon: string;
-    type: "menu" | "submenu" | "checkbox";
-    name: string;
-    hasSeparator: boolean;
-}
-
-export interface IDomContainer {
-    element: JQuery;
-}
+import { IRegisterDirective, ICalcCubeElement } from "./interfaces";
 //#endregion
 
 //#region Logger
